rename platform attach and launch options
